refactor for better structure and logging
