+++ autouri/__main__.py
@@ -1,0 +1,5 @@
+from . import cli
+
+
+if __name__ == '__main__':
+    cli.main()
