+++ autouri/autouri.py
@@ -3,16 +3,12 @@ import os
 from abc import ABC, abstractmethod
 from contextlib import contextmanager
 from filelock import BaseFileLock
-from filelock import logger as logger_filelock
 from typing import Any, Callable, Dict, Optional, Tuple, Union
-
 from .loc_aux import recurse_json, recurse_tsv, recurse_csv
 from .metadata import URIMetadata
 
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s|%(name)s|%(levelname)s| %(message)s')
-logger = logging.getLogger('autouri')
-logger_filelock().setLevel(logging.CRITICAL)
+logger = logging.getLogger(__name__)
 
 
 class AutoURIRecursionError(RuntimeError):
@@ -335,13 +331,14 @@ class URIBase(ABC):
         else:
             return self._get_lock(timeout=timeout, poll_interval=poll_interval)
 
-    def localize_on(self, loc_prefix, recursive=False, make_md5_file=False, depth=0) -> Tuple[str, bool]:
+    def localize_on(self, loc_prefix, recursive=False, make_md5_file=False,
+                    return_flag=False, depth=0) -> Tuple[str, bool]:
         """Wrapper for classmethod localize().
         Localizes self on target directory loc_prefix.
         """
         return AutoURI.localize(
             src_uri=self, recursive=recursive, make_md5_file=make_md5_file,
-            loc_prefix=loc_prefix, return_flag=False, depth=depth)
+            loc_prefix=loc_prefix, return_flag=return_flag, depth=depth)
 
     @abstractmethod
     def _get_lock(self, timeout=None, poll_interval=None) -> BaseFileLock:
