+++ autouri/version.py
@@ -1,0 +1,1 @@
+version = '0.1.2.1'
