+++ bin/autouri
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-
 try:
     import autouri
 except:
@@ -7,7 +6,8 @@ except:
     script_path = os.path.dirname(os.path.realpath(__file__))
     sys.path.append(os.path.join(script_path, '../'))
     import autouri
+from autouri.cli import main
 
 
-autouri.main()
-
+if __name__ == '__main__':
+    main()
